Inline markdownWrapper.hpp
(no need to have it in a separate file)
